+++ src/desktop/remotedirectoryselectionwidget.cpp
@@ -79,7 +79,7 @@ namespace tremotesf
         QStringList currentServerAddTorrentDialogDirectories(Servers::instance()->currentServerAddTorrentDialogDirectories());
         const bool wasEmpty = currentServerAddTorrentDialogDirectories.empty();
 
-        currentServerAddTorrentDialogDirectories.reserve(mRpc->torrents().size() + 1);
+        currentServerAddTorrentDialogDirectories.reserve(static_cast<int>(mRpc->torrents().size() + 1));
         currentServerAddTorrentDialogDirectories.push_back(chopTrailingSeparator(mRpc->serverSettings()->downloadDirectory()));
         for (const auto& torrent : mRpc->torrents()) {
             currentServerAddTorrentDialogDirectories.push_back(chopTrailingSeparator(torrent->downloadDirectory()));
